agents/orchestrator/pyproject.toml
[project]
name = "Orchestrator"
version = "0.1"
description = "Orchestrator agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.128.8",
    "fuzzywuzzy>=0.18.0",
    "httpx>=0.28.1",
    "kagent-langgraph>=0.7.13",
    "langchain>=1.2.0",
    "langchain-core>=1.2.11",
    "langchain-google-genai>=4.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.9",
    "langdetect>=1.0.9",
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langgraph-cli>=0.4.12",
    "langgraph-sdk>=0.3.5",
    "openai>=2.20.0",
    "openinference-instrumentation-langchain>=0.1.59",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-instrumentation-openai>=0.52.3",
    "opentelemetry-sdk>=1.39.1",
    "psycopg[binary,pool]>=3.3.2",

]
